The tabs no longer disappear when scrolling in the extension editor
