+++ newIDE/app/src/EventsBasedBehaviorEditor/EventsBasedBehaviorEditorDialog.js
@@ -4,6 +4,10 @@ import * as React from 'react';
 import Dialog, { DialogPrimaryButton } from '../UI/Dialog';
 import EventsBasedBehaviorEditor from './index';
 import HelpButton from '../UI/HelpButton';
+import { Tabs } from '../UI/Tabs';
+import EventsBasedBehaviorPropertiesEditor from './EventsBasedBehaviorPropertiesEditor';
+
+type TabName = 'configuration' | 'behavior-properties' | 'scene-properties';
 
 type Props = {|
   onApply: () => void,
@@ -14,10 +18,24 @@ type Props = {|
   onRenameSharedProperty: (oldName: string, newName: string) => void,
 |};
 
+type State = {|
+  currentTab: TabName,
+|};
+
 export default class EventsBasedBehaviorEditorDialog extends React.Component<
   Props,
-  {||}
+  State
 > {
+  state = {
+    currentTab: 'configuration',
+  };
+
+  _changeTab = (newTab: TabName) => {
+    this.setState({
+      currentTab: newTab,
+    });
+  };
+
   render() {
     const {
       onApply,
@@ -25,6 +43,7 @@ export default class EventsBasedBehaviorEditorDialog extends React.Component<
       eventsFunctionsExtension,
       project,
     } = this.props;
+    const { currentTab } = this.state;
 
     return (
       <Dialog
@@ -46,22 +65,54 @@ export default class EventsBasedBehaviorEditorDialog extends React.Component<
         open
         onRequestClose={onApply}
         onApply={onApply}
+        fullHeight
+        flexBody
+        fixedContent={
+          <Tabs
+            value={currentTab}
+            onChange={this._changeTab}
+            options={[
+              {
+                value: 'configuration',
+                label: <Trans>Configuration</Trans>,
+              },
+              {
+                value: 'behavior-properties',
+                label: <Trans>Behavior properties</Trans>,
+              },
+              {
+                value: 'scene-properties',
+                label: <Trans>Scene properties</Trans>,
+              },
+            ]}
+          />
+        }
       >
-        <EventsBasedBehaviorEditor
-          project={project}
-          eventsFunctionsExtension={eventsFunctionsExtension}
-          eventsBasedBehavior={eventsBasedBehavior}
-          onTabChanged={
-            () =>
-              this.forceUpdate() /*Force update to ensure dialog is properly positioned*/
-          }
-          onPropertiesUpdated={
-            () =>
-              this.forceUpdate() /*Force update to ensure dialog is properly positioned*/
-          }
-          onRenameProperty={this.props.onRenameProperty}
-          onRenameSharedProperty={this.props.onRenameSharedProperty}
-        />
+        {currentTab === 'configuration' && (
+          <EventsBasedBehaviorEditor
+            project={project}
+            eventsFunctionsExtension={eventsFunctionsExtension}
+            eventsBasedBehavior={eventsBasedBehavior}
+          />
+        )}
+        {currentTab === 'behavior-properties' && (
+          <EventsBasedBehaviorPropertiesEditor
+            allowRequiredBehavior
+            project={project}
+            properties={eventsBasedBehavior.getPropertyDescriptors()}
+            onPropertiesUpdated={() => {}}
+            onRenameProperty={this.props.onRenameProperty}
+            behaviorObjectType={eventsBasedBehavior.getObjectType()}
+          />
+        )}
+        {currentTab === 'shared-properties' && (
+          <EventsBasedBehaviorPropertiesEditor
+            project={project}
+            properties={eventsBasedBehavior.getSharedPropertyDescriptors()}
+            onPropertiesUpdated={() => {}}
+            onRenameProperty={this.props.onRenameSharedProperty}
+          />
+        )}
       </Dialog>
     );
   }
